Refactor ue modal checkbox
