+++ backend/app/app/static/js/map.js
@@ -350,7 +350,7 @@ function api_get_UEs() {
 // 
 function api_get_moving_UEs() {
     
-    var url = app.api_url + '/sim_ue_movement/state-ues';
+    var url = app.api_url + '/ue_movement/state-ues';
 
     $.ajax({
         type: 'GET',
@@ -641,7 +641,7 @@ function fix_points_format( datapoints ) {
 // 
 function api_start_loop( ue ) {
 
-    var url = app.api_url + '/sim_ue_movement/start-loop';
+    var url = app.api_url + '/ue_movement/start-loop';
     var data = {
         "supi": ue.supi
     };
@@ -690,7 +690,7 @@ function api_start_loop( ue ) {
 // 
 function api_stop_loop( ue ) {
 
-    var url = app.api_url + '/sim_ue_movement/stop-loop';
+    var url = app.api_url + '/ue_movement/stop-loop';
     var data = {
         "supi": ue.supi
     };
@@ -755,7 +755,7 @@ function ui_generate_loop_btn_for( ue ) {
 // It also updates the start-all/stop-all button in case all the UEs are moving
 // 
 function ui_set_loop_btn_status_for(ue) {
-    var url = app.api_url + '/sim_ue_movement/state-loop/' + ue.supi;
+    var url = app.api_url + '/ue_movement/state-loop/' + ue.supi;
 
     $.ajax({
         type: 'GET',
