+++ docker-compose.yml
@@ -80,8 +80,8 @@ services:
       - 5672:5672
       - 15672:15672
     volumes:
-      - ~/.docker-conf/rabbitmq/data/:/var/lib/rabbitmq/
-      - ~/.docker-conf/rabbitmq/log/:/var/log/rabbitmq
+      - rabbitmq-data:/var/lib/rabbitmq/
+      - rabbitmq-logs:/var/log/rabbitmq
     env_file:
       - .env
 
@@ -118,4 +118,5 @@ volumes:
   pgadmin-data:
   app-mongo-db-data:
   shared-data:
-  rabbitmq_net:
+  rabbitmq-data:
+  rabbitmq-logs:
